# crates/api/src/routes/events.rs
//! Event REST API endpoints

use crate::{db, error::ApiError, middleware::telegram_auth::AuthenticatedTelegramUser};
use axum::{
    Extension, Json, Router,
    extract::{FromRef, Path, Query, State},
    http::{HeaderMap, StatusCode, header},
    response::{IntoResponse, Response},
    routing::{delete, get, post, put},
};
use chrono::{DateTime, Utc};
use serde::Deserialize;
use sqlx::PgPool;
use televent_core::models::{Event, EventStatus, Timezone};
use typeshare::typeshare;
use utoipa::ToSchema;
use uuid::Uuid;

// Constants for input validation
const MAX_UID_LENGTH: usize = 256;
const MAX_SUMMARY_LENGTH: usize = 256;
const MAX_DESCRIPTION_LENGTH: usize = 10000;
const MAX_LOCATION_LENGTH: usize = 1024;
const MAX_RRULE_LENGTH: usize = 1024;
const MAX_EVENTS_LIMIT: i64 = 1000;

/// Create event request
#[typeshare]
#[derive(Debug, Deserialize, ToSchema)]
pub struct CreateEventRequest {
    /// iCalendar UID (stable across syncs)
    #[schema(example = "unique-uid-123")]
    pub uid: String,
    /// Event summary/title
    #[schema(example = "Team Meeting")]
    pub summary: String,
    /// Detailed description
    pub description: Option<String>,
    /// Event location
    pub location: Option<String>,
    /// Start time (for timed events)
    #[typeshare(serialized_as = "string")]
    pub start: DateTime<Utc>,
    /// End time (for timed events)
    #[typeshare(serialized_as = "string")]
    pub end: DateTime<Utc>,
    /// Whether this is an all-day event
    pub is_all_day: bool,
    /// IANA timezone name
    pub timezone: Timezone,
    /// RFC 5545 recurrence rule
    #[schema(example = "FREQ=WEEKLY;BYDAY=MO")]
    pub rrule: Option<String>,
}

impl CreateEventRequest {
    pub fn validate(&self) -> Result<(), ApiError> {
        if self.uid.len() > MAX_UID_LENGTH {
            return Err(ApiError::BadRequest(format!(
                "UID too long (max {})",
                MAX_UID_LENGTH
            )));
        }
        if self.summary.len() > MAX_SUMMARY_LENGTH {
            return Err(ApiError::BadRequest(format!(
                "Summary too long (max {})",
                MAX_SUMMARY_LENGTH
            )));
        }
        if self
            .description
            .as_ref()
            .is_some_and(|d| d.len() > MAX_DESCRIPTION_LENGTH)
        {
            return Err(ApiError::BadRequest(format!(
                "Description too long (max {})",
                MAX_DESCRIPTION_LENGTH
            )));
        }
        if self
            .location
            .as_ref()
            .is_some_and(|l| l.len() > MAX_LOCATION_LENGTH)
        {
            return Err(ApiError::BadRequest(format!(
                "Location too long (max {})",
                MAX_LOCATION_LENGTH
            )));
        }
        if self
            .rrule
            .as_ref()
            .is_some_and(|r| r.len() > MAX_RRULE_LENGTH)
        {
            return Err(ApiError::BadRequest(format!(
                "RRule too long (max {})",
                MAX_RRULE_LENGTH
            )));
        }
        if self
            .rrule
            .as_ref()
            .is_some_and(|r| r.chars().any(|c| c == '\r' || c == '\n'))
        {
            return Err(ApiError::BadRequest(
                "RRule cannot contain control characters".to_string(),
            ));
        }
        Ok(())
    }
}

/// Update event request
#[typeshare]
#[derive(Debug, Deserialize, ToSchema)]
pub struct UpdateEventRequest {
    pub summary: Option<String>,
    pub description: Option<String>,
    pub location: Option<String>,
    #[typeshare(serialized_as = "string")]
    pub start: Option<DateTime<Utc>>,
    #[typeshare(serialized_as = "string")]
    pub end: Option<DateTime<Utc>>,
    pub is_all_day: Option<bool>,
    pub status: Option<EventStatus>,
    pub rrule: Option<String>,
}

impl UpdateEventRequest {
    pub fn validate(&self) -> Result<(), ApiError> {
        if self
            .summary
            .as_ref()
            .is_some_and(|s| s.len() > MAX_SUMMARY_LENGTH)
        {
            return Err(ApiError::BadRequest(format!(
                "Summary too long (max {})",
                MAX_SUMMARY_LENGTH
            )));
        }
        if self
            .description
            .as_ref()
            .is_some_and(|d| d.len() > MAX_DESCRIPTION_LENGTH)
        {
            return Err(ApiError::BadRequest(format!(
                "Description too long (max {})",
                MAX_DESCRIPTION_LENGTH
            )));
        }
        if self
            .location
            .as_ref()
            .is_some_and(|l| l.len() > MAX_LOCATION_LENGTH)
        {
            return Err(ApiError::BadRequest(format!(
                "Location too long (max {})",
                MAX_LOCATION_LENGTH
            )));
        }
        if self
            .rrule
            .as_ref()
            .is_some_and(|r| r.len() > MAX_RRULE_LENGTH)
        {
            return Err(ApiError::BadRequest(format!(
                "RRule too long (max {})",
                MAX_RRULE_LENGTH
            )));
        }
        if self
            .rrule
            .as_ref()
            .is_some_and(|r| r.chars().any(|c| c == '\r' || c == '\n'))
        {
            return Err(ApiError::BadRequest(
                "RRule cannot contain control characters".to_string(),
            ));
        }
        Ok(())
    }
}

/// List events query parameters
#[typeshare]
#[derive(Debug, Deserialize, ToSchema, utoipa::IntoParams)]
pub struct ListEventsQuery {
    /// Filter events starting after this time
    #[typeshare(serialized_as = "string")]
    pub start: Option<DateTime<Utc>>,
    /// Filter events ending before this time
    #[typeshare(serialized_as = "string")]
    pub end: Option<DateTime<Utc>>,
    /// Maximum number of events to return
    #[schema(default = 100)]
    #[typeshare(serialized_as = "number")]
    pub limit: Option<i64>,
    /// Number of events to skip
    #[schema(default = 0)]
    #[typeshare(serialized_as = "number")]
    pub offset: Option<i64>,
}

/// Event response (same as Event model)
#[typeshare]
pub type EventResponse = Event;

/// Create a new event
#[utoipa::path(
    post,
    path = "/events",
    request_body = CreateEventRequest,
    responses(
        (status = 201, description = "Event created successfully", body = EventResponse),
        (status = 400, description = "Invalid request"),
        (status = 401, description = "Unauthorized")
    ),
    tag = "events",
    security(
        ("telegram_auth" = [])
    )
)]
async fn create_event(
    State(pool): State<PgPool>,
    Extension(auth_user): Extension<AuthenticatedTelegramUser>,
    Json(req): Json<CreateEventRequest>,
) -> Result<Response, ApiError> {
    req.validate()?;
    use crate::db::events::EventTiming;

    let timing = if req.is_all_day {
        EventTiming::AllDay {
            date: req.start.date_naive(),
            end_date: req.end.date_naive(),
        }
    } else {
        EventTiming::Timed {
            start: req.start,
            end: req.end,
        }
    };

    let event = db::events::create_event(
        &pool,
        auth_user.id,
        req.uid,
        req.summary,
        req.description,
        req.location,
        timing,
        req.timezone,
        req.rrule,
    )
    .await?;

    Ok((StatusCode::CREATED, Json(event)).into_response())
}

/// Get event by ID
#[utoipa::path(
    get,
    path = "/events/{id}",
    responses(
        (status = 200, description = "Event details", body = EventResponse),
        (status = 404, description = "Event not found"),
        (status = 401, description = "Unauthorized")
    ),
    params(
        ("id" = Uuid, Path, description = "Event ID")
    ),
    tag = "events",
    security(
        ("telegram_auth" = [])
    )
)]
async fn get_event(
    State(pool): State<PgPool>,
    Extension(auth_user): Extension<AuthenticatedTelegramUser>,
    Path(event_id): Path<Uuid>,
    headers: HeaderMap,
) -> Result<Response, ApiError> {
    let event = db::events::get_event(&pool, auth_user.id, event_id).await?;

    // Check content negotiation
    if let Some(accept) = headers.get(header::ACCEPT)
        && let Ok(accept_str) = accept.to_str()
        && accept_str.contains("text/calendar")
    {
        let attendees = db::events::get_event_attendees(&pool, event_id).await?;
        let ical = super::ical::event_to_ical(&event, &attendees)?;

        return Ok((
            StatusCode::OK,
            [(
                header::CONTENT_TYPE,
                "text/calendar; charset=utf-8".to_string(),
            )],
            ical,
        )
            .into_response());
    }

    Ok(Json(event).into_response())
}

/// List events
#[utoipa::path(
    get,
    path = "/events",
    params(ListEventsQuery),
    responses(
        (status = 200, description = "List of events", body = Vec<EventResponse>),
        (status = 401, description = "Unauthorized")
    ),
    tag = "events",
    security(
        ("telegram_auth" = [])
    )
)]
async fn list_events(
    State(pool): State<PgPool>,
    Extension(auth_user): Extension<AuthenticatedTelegramUser>,
    Query(query): Query<ListEventsQuery>,
) -> Result<Json<Vec<EventResponse>>, ApiError> {
    // Default limit to 100 to prevent OOM
    let limit = query.limit.unwrap_or(100).clamp(1, MAX_EVENTS_LIMIT);
    let offset = query.offset.unwrap_or(0);

    let events = db::events::list_events(
        &pool,
        auth_user.id,
        query.start,
        query.end,
        Some(limit),
        Some(offset),
    )
    .await?;
    Ok(Json(events))
}

/// Update event
#[utoipa::path(
    put,
    path = "/events/{id}",
    request_body = UpdateEventRequest,
    responses(
        (status = 200, description = "Event updated successfully", body = EventResponse),
        (status = 404, description = "Event not found"),
        (status = 400, description = "Invalid request"),
        (status = 401, description = "Unauthorized")
    ),
    params(
        ("id" = Uuid, Path, description = "Event ID")
    ),
    tag = "events",
    security(
        ("telegram_auth" = [])
    )
)]
async fn update_event(
    State(pool): State<PgPool>,
    Extension(auth_user): Extension<AuthenticatedTelegramUser>,
    Path(event_id): Path<Uuid>,
    Json(req): Json<UpdateEventRequest>,
) -> Result<Json<EventResponse>, ApiError> {
    req.validate()?;

    // Get current event to determine how to set date fields
    let current = db::events::get_event(&pool, auth_user.id, event_id).await?;
    let is_all_day = req.is_all_day.unwrap_or(current.is_all_day);

    // Determine start_date/end_date based on is_all_day
    let (start, end, start_date, end_date) = if is_all_day {
        // For all-day events, use date fields
        let sd = req.start.map(|s| s.date_naive()).or(current.start_date);
        let ed = req.end.map(|e| e.date_naive()).or(current.end_date);
        (None, None, sd, ed)
    } else {
        // For timed events, use time fields
        let s = req.start.or(current.start);
        let e = req.end.or(current.end);
        (s, e, None, None)
    };

    let mut conn = pool
        .acquire()
        .await
        .map_err(|e| ApiError::Internal(format!("Database error: {}", e)))?;

    let event = db::events::update_event(
        &mut conn,
        auth_user.id,
        event_id,
        req.summary,
        req.description,
        req.location,
        start,
        end,
        start_date,
        end_date,
        req.is_all_day,
        req.status,
        req.rrule,
    )
    .await?;
    Ok(Json(event))
}

/// Delete event
#[utoipa::path(
    delete,
    path = "/events/{id}",
    responses(
        (status = 201, description = "Event deleted successfully"),
        (status = 404, description = "Event not found"),
        (status = 401, description = "Unauthorized")
    ),
    params(
        ("id" = Uuid, Path, description = "Event ID")
    ),
    tag = "events",
    security(
        ("telegram_auth" = [])
    )
)]
async fn delete_event_handler(
    State(pool): State<PgPool>,
    Extension(auth_user): Extension<AuthenticatedTelegramUser>,
    Path(event_id): Path<Uuid>,
) -> Result<StatusCode, ApiError> {
    db::events::delete_event(&pool, auth_user.id, event_id).await?;
    Ok(StatusCode::NO_CONTENT)
}

/// Event routes
pub fn routes<S>() -> Router<S>
where
    S: Clone + Send + Sync + 'static,
    PgPool: FromRef<S>,
{
    Router::new()
        .route("/events", post(create_event))
        .route("/events", get(list_events))
        .route("/events/{id}", get(get_event))
        .route("/events/{id}", put(update_event))
        .route("/events/{id}", delete(delete_event_handler))
}

#[cfg(test)]
mod tests {
    use super::*;

    #[test]
    fn test_create_event_request_deserialization() {
        let json = r#"{
            "uid": "test-uid",
            "summary": "Test Event",
            "description": "Test Description",
            "location": "Test Location",
            "start": "2026-01-18T10:00:00Z",
            "end": "2026-01-18T11:00:00Z",
            "is_all_day": false,
            "timezone": "UTC",
            "rrule": null
        }"#;

        let req: CreateEventRequest = serde_json::from_str(json).unwrap();
        assert_eq!(req.summary, "Test Event");
        assert_eq!(req.timezone.to_string(), "UTC");
        assert_eq!(req.description, Some("Test Description".to_string()));
        assert!(req.rrule.is_none());
    }

    #[test]
    fn test_update_event_request_partial() {
        let json = r#"{
            "summary": "Updated Summary"
        }"#;

        let req: UpdateEventRequest = serde_json::from_str(json).unwrap();
        assert_eq!(req.summary, Some("Updated Summary".to_string()));
        assert!(req.description.is_none());
        assert!(req.start.is_none());
    }

    #[test]
    fn test_list_events_query_deserialization() {
        let json = r#"{
            "start": "2026-01-18T10:00:00Z",
            "limit": 50,
            "offset": 10
        }"#;

        let query: ListEventsQuery = serde_json::from_str(json).unwrap();
        assert_eq!(query.limit, Some(50));
        assert_eq!(query.offset, Some(10));
        assert!(query.start.is_some());
        assert!(query.end.is_none());
    }

    #[test]
    fn test_list_events_query_default_values() {
        let json = r#"{}"#;

        let query: ListEventsQuery = serde_json::from_str(json).unwrap();
        assert!(query.limit.is_none());
        assert!(query.offset.is_none());
    }

    #[test]
    fn test_create_event_validation_success() {
        let req = CreateEventRequest {
            uid: "valid-uid".to_string(),
            summary: "Valid Summary".to_string(),
            description: Some("Valid Description".to_string()),
            location: Some("Valid Location".to_string()),
            start: Utc::now(),
            end: Utc::now(),
            is_all_day: false,
            timezone: televent_core::models::Timezone::default(),
            rrule: None,
        };
        assert!(req.validate().is_ok());
    }

    #[test]
    fn test_create_event_validation_too_long() {
        let req = CreateEventRequest {
            uid: "a".repeat(MAX_UID_LENGTH + 1),
            summary: "Valid Summary".to_string(),
            description: None,
            location: None,
            start: Utc::now(),
            end: Utc::now(),
            is_all_day: false,
            timezone: televent_core::models::Timezone::default(),
            rrule: None,
        };
        assert!(req.validate().is_err());

        let req = CreateEventRequest {
            uid: "valid-uid".to_string(),
            summary: "a".repeat(MAX_SUMMARY_LENGTH + 1),
            description: None,
            location: None,
            start: Utc::now(),
            end: Utc::now(),
            is_all_day: false,
            timezone: televent_core::models::Timezone::default(),
            rrule: None,
        };
        assert!(req.validate().is_err());
    }

    #[test]
    fn test_update_event_validation() {
        let req = UpdateEventRequest {
            summary: Some("a".repeat(MAX_SUMMARY_LENGTH + 1)),
            description: None,
            location: None,
            start: None,
            end: None,
            is_all_day: None,
            status: None,
            rrule: None,
        };
        assert!(req.validate().is_err());

        let req = UpdateEventRequest {
            summary: Some("Valid".to_string()),
            description: None,
            location: None,
            start: None,
            end: None,
            is_all_day: None,
            status: None,
            rrule: None,
        };
        assert!(req.validate().is_ok());
    }

    #[test]
    fn test_create_event_validation_rrule_injection() {
        let req = CreateEventRequest {
            uid: "valid-uid".to_string(),
            summary: "Valid Summary".to_string(),
            description: None,
            location: None,
            start: Utc::now(),
            end: Utc::now(),
            is_all_day: false,
            timezone: televent_core::models::Timezone::default(),
            rrule: Some("FREQ=DAILY\r\nATTENDEE:EVIL".to_string()),
        };
        assert!(req.validate().is_err());
    }
}