# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "contact": {
      "name": "kirill denisov"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List events",
        "operationId": "list_events",
        "parameters": [
          {
            "name": "start",
            "in": "path",
            "description": "Filter events starting after this time",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "path",
            "description": "Filter events ending before this time",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "path",
            "description": "Maximum number of events to return",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "path",
            "description": "Number of events to skip",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "telegram_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Create a new event",
        "operationId": "create_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "telegram_auth": []
          }
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get event by ID",
        "operationId": "get_event",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Event ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "telegram_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "events"
        ],
        "summary": "Update event",
        "operationId": "update_event",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Event ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "telegram_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "events"
        ],
        "summary": "Delete event",
        "operationId": "delete_event_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Event ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Event deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "telegram_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEventRequest": {
        "type": "object",
        "description": "Create event request",
        "required": [
          "uid",
          "summary",
          "start",
          "end",
          "is_all_day",
          "timezone"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Detailed description"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "description": "End time (for timed events)"
          },
          "is_all_day": {
            "type": "boolean",
            "description": "Whether this is an all-day event"
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "description": "Event location"
          },
          "rrule": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 5545 recurrence rule",
            "example": "FREQ=WEEKLY;BYDAY=MO"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Start time (for timed events)"
          },
          "summary": {
            "type": "string",
            "description": "Event summary/title",
            "example": "Team Meeting"
          },
          "timezone": {
            "$ref": "#/components/schemas/Timezone",
            "description": "IANA timezone name"
          },
          "uid": {
            "type": "string",
            "description": "iCalendar UID (stable across syncs)",
            "example": "unique-uid-123"
          }
        }
      },
...
      "Event": {
        "type": "object",
        "description": "Event entity",
        "required": [
          "id",
          "user_id",
          "uid",
          "summary",
          "is_all_day",
          "status",
          "timezone",
          "version",
          "etag",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "end": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "end_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "etag": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_all_day": {
            "type": "boolean"
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "rrule": {
            "type": [
              "string",
              "null"
            ]
          },
          "start": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "start_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "summary": {
            "type": "string"
          },
          "timezone": {
            "$ref": "#/components/schemas/Timezone"
          },
          "uid": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "string",
            "description": "Owner's user ID (telegram_id)"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "EventStatus": {
        "type": "string",
        "description": "Event status enumeration",
        "enum": [
          "Confirmed",
          "Tentative",
          "Cancelled"
        ]
      },
...
      "Timezone": {
        "type": "string",
        "description": "Timezone newtype wrapping chrono_tz::Tz with SQLx and Serde support\n\nStored in database as TEXT (IANA timezone name like \"America/New_York\")",
        "example": "UTC"
      },
      "UpdateEventRequest": {
        "type": "object",
        "description": "Update event request",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "end": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "is_all_day": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "rrule": {
            "type": [
              "string",
              "null"
            ]
          },
          "start": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventStatus"
              }
            ]
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "telegram_auth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-telegram-init-data"
      }
    }
  },
...
}